Make the speculation id an int and allow for more bits to be set in the HotSpot implementation (which currently silently discards unsupported high bit).

+++ graal/com.oracle.graal.api.meta/src/com/oracle/graal/api/meta/MetaAccessProvider.java
@@ -74,7 +74,7 @@ public interface MetaAccessProvider {
      * 
      * @return the encoded value as an integer
      */
-    Constant encodeDeoptActionAndReason(DeoptimizationAction action, DeoptimizationReason reason, short speculationId);
+    Constant encodeDeoptActionAndReason(DeoptimizationAction action, DeoptimizationReason reason, int speculationId);
 
     DeoptimizationReason decodeDeoptReason(Constant constant);
